refactor: GlobalConfig 패키지 변경
잘못 구성된 패키지 설정 수정

- 기존: http.config.GlobalConfig
- 변경: config.GlobalConfig
